+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 - Clean up docs for OME-TIFF creation.
 - Remove OMETIFF_LOADING.md docs.
 - Upgrade deck.gl to 8.4.0-alpha.2.
-- Remove `scale` and `translate` from `ImageLayer` in favor of `modelMatrix`.
+- Remove `scale` and `translate` from `ImageLayer` and its usages in favor of `modelMatrix`.
 
 ## 0.6.0